Abstract out parent Gecko Preferences activity.
