+++ mobile/android/base/GeckoPreferences.java
@@ -7,7 +7,6 @@ package org.mozilla.gecko;
 
 import org.mozilla.gecko.background.announcements.AnnouncementsConstants;
 import org.mozilla.gecko.util.GeckoEventListener;
-import org.mozilla.gecko.util.ThreadUtils;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -17,15 +16,9 @@ import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
-import android.content.SharedPreferences;
-import android.os.Build;
 import android.os.Bundle;
-import android.preference.CheckBoxPreference;
-import android.preference.EditTextPreference;
 import android.preference.ListPreference;
 import android.preference.Preference;
-import android.preference.Preference.OnPreferenceChangeListener;
-import android.preference.PreferenceActivity;
 import android.preference.PreferenceGroup;
 import android.preference.PreferenceManager;
 import android.preference.PreferenceScreen;
@@ -34,24 +27,14 @@ import android.text.InputType;
 import android.text.TextUtils;
 import android.text.TextWatcher;
 import android.util.Log;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.LinearLayout;
-import android.widget.Toast;
-
-import java.util.ArrayList;
 
 public class GeckoPreferences
-    extends PreferenceActivity
-    implements OnPreferenceChangeListener, GeckoEventListener, GeckoActivityStatus
+    extends GeckoPreferencesActivity
 {
     private static final String LOGTAG = "GeckoPreferences";
-
-    private ArrayList<String> mPreferencesList;
-    private PreferenceScreen mPreferenceScreen;
-    private static boolean sIsCharEncodingEnabled = false;
-    private static final String NON_PREF_PREFIX = "android.not_a_preference.";
 
     // These match keys in resources/xml/preferences.xml.in.
     private static String PREFS_ANNOUNCEMENTS_ENABLED = NON_PREF_PREFIX + "privacy.announcements.enabled";
@@ -59,134 +42,25 @@ public class GeckoPreferences
     private static String PREFS_CATEGORY_PRIVACY = "category_privacy";
     private static String PREFS_MENU_CHAR_ENCODING = "browser.menu.showCharacterEncoding";
     private static String PREFS_MP_ENABLED = "privacy.masterpassword.enabled";
-    private static String PREFS_TELEMETRY_ENABLED = "toolkit.telemetry.enabled";
-    private static String PREFS_TELEMETRY_ENABLED_PRERELEASE = "toolkit.telemetry.enabledPreRelease";
     private static String PREFS_UPDATER_AUTODOWNLOAD = "app.update.autodownload";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        addPreferencesFromResource(R.xml.preferences);
-        registerEventListener("Sanitize:Finished");
-
-        if (Build.VERSION.SDK_INT >= 14)
-            getActionBar().setHomeButtonEnabled(true);
 
         mPreferenceScreen = getPreferenceScreen();
         if (!AppConstants.MOZ_UPDATER) {
             ((PreferenceGroup) mPreferenceScreen.findPreference(PREFS_CATEGORY_GENERAL))
                     .removePreference(findPreference(PREFS_UPDATER_AUTODOWNLOAD));
         }
-
-        Preference telemetryPref = findPreference(PREFS_TELEMETRY_ENABLED);
-        if (AppConstants.MOZ_TELEMETRY_REPORTING) {
-            if (AppConstants.MOZ_TELEMETRY_ON_BY_DEFAULT) {
-                telemetryPref.setKey(PREFS_TELEMETRY_ENABLED_PRERELEASE);
-            }
-        } else {
-            ((PreferenceGroup) mPreferenceScreen.findPreference(PREFS_CATEGORY_PRIVACY))
-                    .removePreference(telemetryPref);
-        }
-
     }
 
-    @Override
-    public void onWindowFocusChanged(boolean hasFocus) {
-        if (!hasFocus)
-            return;
-
-        mPreferencesList = new ArrayList<String>();
-        initGroups(mPreferenceScreen);
-        initValues();
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        unregisterEventListener("Sanitize:Finished");
-    }
-
-    @Override
-    public void onPause() {
-        super.onPause();
-
-        if (getApplication() instanceof GeckoApplication) {
-            ((GeckoApplication) getApplication()).onActivityPause(this);
-        }
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-
-        if (getApplication() instanceof GeckoApplication) {
-            ((GeckoApplication) getApplication()).onActivityResume(this);
-        }
-    }
-
-    @Override
-    public void handleMessage(String event, JSONObject message) {
-        try {
-            if (event.equals("Sanitize:Finished")) {
-                boolean success = message.getBoolean("success");
-                final int stringRes = success ? R.string.private_data_success : R.string.private_data_fail;
-                final Context context = this;
-                ThreadUtils.postToUiThread(new Runnable () {
-                    @Override
-                    public void run() {
-                        Toast.makeText(context, stringRes, Toast.LENGTH_SHORT).show();
-                    }
-                });
-            }
-        } catch (Exception e) {
-            Log.e(LOGTAG, "Exception handling message \"" + event + "\":", e);
-        }
-    }
-
-    private void initGroups(PreferenceGroup preferences) {
-        final int count = preferences.getPreferenceCount();
-        for (int i = 0; i < count; i++) {
-            Preference pref = preferences.getPreference(i);
-            if (pref instanceof PreferenceGroup)
-                initGroups((PreferenceGroup)pref);
-            else {
-                pref.setOnPreferenceChangeListener(this);
-
-                // Some Preference UI elements are not actually preferences,
-                // but they require a key to work correctly. For example,
-                // "Clear private data" requires a key for its state to be
-                // saved when the orientation changes. It uses the
-                // "android.not_a_preference.privacy.clear" key - which doesn't
-                // exist in Gecko - to satisfy this requirement.
-                String key = pref.getKey();
-                if (key != null && !key.startsWith(NON_PREF_PREFIX)) {
-                    mPreferencesList.add(pref.getKey());
-                }
-            }
-        }
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case android.R.id.home:
-                finish();
-                return true;
-        }
-
-        return super.onOptionsItemSelected(item);
+    protected int getPreferencesResource() {
+        return R.xml.preferences;
     }
 
     final private int DIALOG_CREATE_MASTER_PASSWORD = 0;
     final private int DIALOG_REMOVE_MASTER_PASSWORD = 1;
-
-    public static void setCharEncodingState(boolean enabled) {
-        sIsCharEncodingEnabled = enabled;
-    }
-
-    public static boolean getCharEncodingState() {
-        return sIsCharEncodingEnabled;
-    }
 
     /**
      * Broadcast an intent with <code>pref</code>, <code>branch</code>, and
@@ -224,23 +98,6 @@ public class GeckoPreferences
     public static void broadcastAnnouncementsPref(final Context context) {
         final boolean value = getBooleanPref(context, PREFS_ANNOUNCEMENTS_ENABLED, true);
         broadcastAnnouncementsPref(context, value);
-    }
-
-    /**
-     * Return the value of the named preference in the default preferences file.
-     *
-     * This corresponds to the storage that backs preferences.xml.
-     * @param context a <code>Context</code>; the
-     *                <code>PreferenceActivity</code> will suffice, but this
-     *                method is intended to be called from other contexts
-     *                within the application, not just this <code>Activity</code>.
-     * @param name    the name of the preference to retrieve.
-     * @param def     the default value to return if the preference is not present.
-     * @return        the value of the preference, or the default.
-     */
-    public static boolean getBooleanPref(final Context context, final String name, boolean def) {
-        final SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
-        return prefs.getBoolean(name, def);
     }
 
     @Override
@@ -429,85 +286,6 @@ public class GeckoPreferences
             default:
                 return null;
         }
-
         return dialog;
-    }
-
-    // Initialize preferences by requesting the preference values from Gecko
-    private void initValues() {
-        JSONArray jsonPrefs = new JSONArray(mPreferencesList);
-        PrefsHelper.getPrefs(jsonPrefs, new PrefsHelper.PrefHandlerBase() {
-            private Preference getField(String prefName) {
-                return (mPreferenceScreen == null ? null : mPreferenceScreen.findPreference(prefName));
-            }
-
-            @Override public void prefValue(String prefName, final boolean value) {
-                final Preference pref = getField(prefName);
-                if (pref instanceof CheckBoxPreference) {
-                    ThreadUtils.postToUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            if (((CheckBoxPreference)pref).isChecked() != value)
-                                ((CheckBoxPreference)pref).setChecked(value);
-                        }
-                    });
-                }
-            }
-
-            @Override public void prefValue(String prefName, final String value) {
-                final Preference pref = getField(prefName);
-                if (pref instanceof EditTextPreference) {
-                    ThreadUtils.postToUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            ((EditTextPreference)pref).setText(value);
-                        }
-                    });
-                } else if (pref instanceof ListPreference) {
-                    ThreadUtils.postToUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            ((ListPreference)pref).setValue(value);
-                            // Set the summary string to the current entry
-                            CharSequence selectedEntry = ((ListPreference)pref).getEntry();
-                            ((ListPreference)pref).setSummary(selectedEntry);
-                        }
-                    });
-                } else if (pref instanceof FontSizePreference) {
-                    final FontSizePreference fontSizePref = (FontSizePreference) pref;
-                    fontSizePref.setSavedFontSize(value);
-                    final String fontSizeName = fontSizePref.getSavedFontSizeName();
-                    ThreadUtils.postToUiThread(new Runnable() {
-                        @Override
-                        public void run() {
-                            fontSizePref.setSummary(fontSizeName); // Ex: "Small".
-                        }
-                    });
-                }
-            }
-
-            @Override public void finish() {
-                // enable all preferences once we have them from gecko
-                ThreadUtils.postToUiThread(new Runnable() {
-                    @Override
-                    public void run() {
-                        mPreferenceScreen.setEnabled(true);
-                    }
-                });
-            }
-        });
-    }
-
-    private void registerEventListener(String event) {
-        GeckoAppShell.getEventDispatcher().registerEventListener(event, this);
-    }
-
-    private void unregisterEventListener(String event) {
-        GeckoAppShell.getEventDispatcher().unregisterEventListener(event, this);
-    }
-
-    @Override
-    public boolean isGeckoActivityOpened() {
-        return false;
     }
 }
